+++ frontend/test-lk.js
@@ -1,0 +1,2 @@
+import { VideoTrack } from "@livekit/components-react";
+console.log(VideoTrack);